feat(volunteer): update profile screen
- moved logout in the header on main profile screen per
  new `useLogoutHeader` hook, that can be reused later
- updated icon for navigating to the profile screen to
  a person and used the lock for logout

HDVT-17
