+++ src/helpers/parser/volunteer/profileData.js
@@ -1,8 +1,3 @@
-import { Alert } from 'react-native';
-
-import { storeVolunteerAuthToken } from '../../volunteerHelper';
-import { ScreenName } from '../../../types';
-
 export const myProfile = () => [
   {
     id: 1,
@@ -14,28 +9,6 @@ export const myProfile = () => [
   },
   {
     id: 3,
-    title: 'Logout',
-    onPress: async (navigation) => {
-      Alert.alert(
-        'Abmelden',
-        'Bist du sicher, dass du dich abmelden möchtest?',
-        [
-          {
-            text: 'Abbrechen'
-          },
-          {
-            text: 'Ja, abmelden',
-            style: 'destructive',
-            onPress: async () => {
-              await storeVolunteerAuthToken();
-              navigation?.navigate(ScreenName.VolunteerHome, {
-                refreshUser: new Date().valueOf()
-              });
-            }
-          }
-        ],
-        { cancelable: false }
-      );
-    }
+    title: 'Gespeichert'
   }
 ];
