+++ src/helpers/volunteerHelper.ts
@@ -2,7 +2,7 @@ import * as SecureStore from 'expo-secure-store';
 
 const VOLUNTEER_AUTH_TOKEN = 'VOLUNTEER_AUTH_TOKEN';
 
-export const storeVolunteerAuthToken = (authToken: string) => {
+export const storeVolunteerAuthToken = (authToken?: string) => {
   if (authToken) {
     SecureStore.setItemAsync(VOLUNTEER_AUTH_TOKEN, authToken);
   } else {
